+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next [????.??.??]
+* Add `genericLiftTyped` and `genericLiftTypedWithPkg`.
+
 ### 0.1.2
 * Add a case for `V1` (i.e., empty data types), which diverges in the way
   you'd expect.